Tested and working!
First attempt seemed to work but did not result in sandwich at my
location. I went to Jimmy John's to figure out what went wrong. For some
reason the phone number shows up as 000-000-0000, but I couldn't figure
it out. Instead, I put the number in the DeliveryInstructions and that
seemed to work.